+++ src/extension.ts
@@ -1,6 +1,7 @@
 import * as vscode from 'vscode';
 import { closeAllExceptMdCommand } from './command/closeAllExceptMd';
 import { convertToWorkspaceCommand } from './command/convertToWorkspace';
+import { activateFluentIconPreview } from './command/fluentIconPreview';
 import { newMemoCommand } from './command/newMemo';
 import { selectMemoDirCommand } from './command/selectMemoDir';
 import { selectWorkspaceDirCommand } from './command/selectWorkspaceDir';
@@ -38,4 +39,6 @@ export function activate(context: vscode.ExtensionContext) {
 	);
 
 	context.subscriptions.push(selectMemoDirCmd, selectWorkspaceDirCmd, newMemoCmd, newMemoFromSelectionCmd, convertToWorkspaceCmd, closeAllExceptMdCmd);
+
+	activateFluentIconPreview(context);
 }